- fixed some problems with updating issue, EditIssueAction (wrong notification, not updated properties like severity, status, resolution...) - added dependency to newser hibernate for increasing stability (hibernate 3.2.4ga) - Minor Fixes to DAO (some Long-retuns for count-values), debugging

+++ src/main/java/org/itracker/persistence/dao/IssueAttachmentDAO.java
@@ -37,7 +37,7 @@ public interface IssueAttachmentDAO extends BaseDAO<IssueAttachment> {
      * @return count of all issue attachments in system
      */ 
     
-    Integer countAll();
+    Long countAll();
     /**
      * Calculates the total Size of attachments in the system. 
      *
